+++ lib/Braintree/WebhookNotification.php
@@ -31,6 +31,7 @@ class WebhookNotification extends Base
     const IDEAL_PAYMENT_COMPLETE = 'ideal_payment_complete';
     const IDEAL_PAYMENT_FAILED = 'ideal_payment_failed';
     const GRANTED_PAYMENT_INSTRUMENT_UPDATE = 'granted_payment_instrument_update';
+    const LOCAL_PAYMENT_COMPLETED = "local_payment_completed";
 
     public static function parse($signature, $payload) {
         return Configuration::gateway()->webhookNotification()->parse($signature, $payload);
@@ -107,6 +108,10 @@ class WebhookNotification extends Base
 
         if (isset($wrapperNode['grantedPaymentInstrumentUpdate'])) {
             $this->_set('grantedPaymentInstrumentUpdate', GrantedPaymentInstrumentUpdate::factory($wrapperNode['grantedPaymentInstrumentUpdate']));
+        }
+
+        if (isset($wrapperNode['localPayment'])) {
+            $this->_set('localPayment', LocalPaymentCompleted::factory($wrapperNode['localPayment']));
         }
 
         if (isset($wrapperNode['errors'])) {